+++ tags/categoryList.go
@@ -106,7 +106,7 @@ func (node *tagCategoryListNode) Execute(ctx *pongo2.ExecutionContext, writer po
 		categoryList[i].Thumb = categoryList[i].GetThumb(currentSite.PluginStorage.StorageUrl, currentSite.Content.DefaultThumb)
 		categoryList[i].IsCurrent = false
 		if webOk {
-			if (webInfo.PageName == "archiveList" || webInfo.PageName == "archiveDetail") && categoryList[i].Id == webInfo.NavBar {
+			if (webInfo.PageName == "archiveList" || webInfo.PageName == "archiveDetail") && int64(categoryList[i].Id) == webInfo.NavBar {
 				categoryList[i].IsCurrent = true
 			}
 		}
